Setup versioning of the PVP itself
Originally suggested in

https://groups.google.com/forum/#!topic/haskell-core-libraries/xkJT8HEh7tM

This implements the suggestion from #19, i.e.

> So I'd actually propose we go with 2-part version numbers; and we
> declare the current version to be labeled as "1.0".
>
> We can then have "http://pvp.haskell.org/v1.0" be the permalink to
> version 1.0, and have http://pvp.haskell.org/ redirect to the latest
> current version, i.e. pvp.haskell.org/v1.0; once there is say v1.1,
> we'd make sure that all non-latest versions have a warning at the top
> denoting this isn't the latest pvp-spec version.

The versioning of the PVP is intended to denote behavioral/semantic
changes to the spec; typos or minor clarifications which don't change
the meaning of the PVP contract therefore don't require allocating a
new version.

This also sets up a (clarly marked) draft version v1.1 to serve as
proof of concept

resolves #19